+++ src/hash.h
@@ -4,13 +4,14 @@
 // This defines some simple hash functions
 
 #include <stddef.h>
+#include <stdint.h>
 
-size_t hashBytes(const void* mem, size_t size);
+uint64_t hashBytes(const void* mem, size_t size);
 
-size_t hashInt(size_t value);
+uint64_t hashInt(uint64_t value);
 
-size_t hashString(const char* s);
+uint64_t hashString(const char* s);
 
-size_t hashCombine(size_t first, size_t second);
+uint64_t hashCombine(uint64_t first, uint64_t second);
 
 #endif
